Improve ticker balances update logic
The previous mechanism updated all the tickers and USD
balances at once on the server side every X seconds and sent the all the balances
to be processed on the 'balances:update' event channel on the client side.
This meant that the client wasn't able to request specific ticker/account balances when needed.

With this change we create a client side method named `requestCoinBalance` which we call per ticker
as needed and process the responses in the 'balances:singleTicker' channel (see index.html:157).
This method is also used to fetch the account USD balance and total account value.

This new design also allows us to fix the 'Add ticker' issue mentioned in the issue tracker (see #2). See
index.html:266

Finally we also improve the real time ticker balances display by updating balances when new order fills are recieved
on the 'orders:fill' channel